+++ src/components/Layout/AffiliationContainer/index.tsx
@@ -3,24 +3,13 @@
 import classNames from "classnames/bind";
 import styles from "./index.module.scss";
 import AffiliationLabel from "@/components/common/AffiliationLabel";
-import { useGetMyInfo } from "@/hooks/tanstack-query/common/my-info/useGetMyInfo";
 import Skeleton from "@/components/common/Skeleton";
-import { useRouter } from "next/navigation";
-import { ROUTE } from "@/constants/routes";
-import { useEffect } from "react";
+import { useStudentMyInfo } from "@/hooks/common/useStudentMyInfo";
 
 const cn = classNames.bind(styles);
 
 export default function AffiliationContainer() {
-  const router = useRouter();
-  const { data, isPending, isError, error } = useGetMyInfo();
-
-  useEffect(() => {
-    if (isError && error.response.status === 401) {
-      alert("로그인 후 이용해주세요.");
-      router.push(ROUTE.STUDENT.MAIN);
-    }
-  }, [isError, error, router]);
+  const { data, isError, isPending } = useStudentMyInfo();
 
   if (isPending) {
     return <Skeleton className={cn("skeleton")} />;
@@ -32,7 +21,7 @@ export default function AffiliationContainer() {
 
   return (
     <div className={cn("container")}>
-      <AffiliationLabel affiliation={data.affiliation} labelClassName={cn("label")} />
+      <AffiliationLabel affiliation={`${data?.affiliation}`} labelClassName={cn("label")} />
     </div>
   );
 }
